Predictive-Maintenance-Agent/pyproject.toml
[project]
name = "karley-scheduling-agent"
version = "0.1.0"
description = "Karley's scheduling agent for the A2A project."
requires-python = ">=3.13, <3.14"
dependencies = [
    "a2a-sdk==0.2.5",
    "google-adk==1.2.1",
    "google-api-core==2.25.0",
    "google-api-python-client==2.171.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform==1.96.0",
    "google-cloud-appengine-logging==1.6.1",
    "google-cloud-audit-log==0.3.2",
    "google-cloud-bigquery==3.34.0",
    "google-cloud-core==2.4.3",
    "google-cloud-logging==3.12.1",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-secret-manager==2.24.0",
    "google-cloud-speech==2.32.0",
    "google-cloud-storage==2.19.0",
    "google-cloud-trace==1.16.1",
    "google-crc32c==1.7.1",
    "google-genai==1.19.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "numpy==2.2.6",
    "pandas==3.0.0",
    "torch==2.10.0",
    "uvicorn==0.34.3"
] 
